feat: allow users to quit from multiselect menu

+++ src/core/lib/downloaders/custom_downloader.py
@@ -49,17 +49,28 @@ class CustomDownloader(VideoDownloaderProtocol[list[VideoDownloadResult]]):
 
         chosen_streams = pick_from_streams(await youtube_video.streams())
 
-        logger.info("suitable stream(s) successfully chosen")
         logger.debug(
             "video_id=%s stream_itags=%s should_convert=%s should_skip_existing_files=%s custom_file_extension=%s delete_temporary_files=%s put_custom_streams_in_folder=%s",
             youtube_video.video_id,
-            map(lambda x: x.itag, chosen_streams),
+            tuple(map(lambda x: x.itag, chosen_streams)),
             self._configuration["download_behavior_configuration"]["convert_custom_downloads"], 
             self._configuration["download_behavior_configuration"]["skip_existing_files"],
             self._configuration["download_behavior_configuration"]["convert_custom_downloads_to"],
             self._configuration["download_behavior_configuration"]["automatically_delete_temporary_files_after_download"],
             put_in_custom_folder
         )
+
+        if len(chosen_streams) == 0:
+            logger.info("no streams chosen")
+            
+            return [{
+                "success": False,
+                "by_user_action": True,
+                "youtube_video": youtube_video,
+                "message": "was canceled by the user."
+            }]
+
+        logger.info("suitable stream(s) successfully chosen")
 
         if put_in_custom_folder:
             download_directory = safe_join_directory(
@@ -70,7 +81,7 @@ class CustomDownloader(VideoDownloaderProtocol[list[VideoDownloadResult]]):
             download_directory.mkdir(parents=True, exist_ok=True)
 
         for stream in chosen_streams:
-            result = await self._inner_download(
+            result = await self._custom_stream_download(
                 stream,
                 download_directory,
                 filename_prefix
@@ -78,7 +89,7 @@ class CustomDownloader(VideoDownloaderProtocol[list[VideoDownloadResult]]):
 
             results.append(result)
 
-        if all(map(lambda r: not r["success"], results)):
+        if all(map(lambda r: not r["success"], results)) and put_in_custom_folder:
             download_directory.rmdir()
 
         if delete_temporary_files:
@@ -90,7 +101,7 @@ class CustomDownloader(VideoDownloaderProtocol[list[VideoDownloadResult]]):
         return results
 
 
-    async def _inner_download(
+    async def _custom_stream_download(
         self, 
         stream: Stream,
         download_directory: Path,
@@ -136,9 +147,9 @@ class CustomDownloader(VideoDownloaderProtocol[list[VideoDownloadResult]]):
 
             return {
                 "success": False,
+                "by_user_action": False,
                 "youtube_video": youtube_video,
-                "download_path": None,
-                "error_message": str(exception)
+                "message": str(exception)
             }
 
         if converted_file_path.exists() and should_skip_existing_files:
@@ -150,9 +161,9 @@ class CustomDownloader(VideoDownloaderProtocol[list[VideoDownloadResult]]):
 
             return {
                 "success": False,
+                "by_user_action": False,
                 "youtube_video": youtube_video,
-                "download_path": None,
-                "error_message": str.format(ALREADY_EXISTS_AT_PATH_ERROR_MESSAGE, path=converted_file_path)
+                "message": str.format(ALREADY_EXISTS_AT_PATH_ERROR_MESSAGE, path=converted_file_path)
             }
 
         logger.info("beginning stream download")
@@ -203,6 +214,5 @@ class CustomDownloader(VideoDownloaderProtocol[list[VideoDownloadResult]]):
         return {
             "success": True,
             "youtube_video": youtube_video,
-            "download_path": converted_file_path,
-            "error_message": None
+            "download_path": converted_file_path
         }
